Safari-611.1.10/LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-single-value-in-head-expected.txt


FAIL Meta color-scheme in head applies. assert_equals: expected (string) "dark" but got (undefined) undefined
